VidSketch/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
accelerate=1.0.1=pypi_0
affine=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.11=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.5.2=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
ca-certificates=2024.12.14=hbcca054_0
cachetools=5.5.0=pypi_0
cairocffi=1.7.1=pypi_0
cairosvg=2.7.1=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
clip=0.2.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
controlnet-aux=0.0.9=pypi_0
cssselect2=0.7.0=pypi_0
cssutils=2.11.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=4.4.2=pypi_0
decord=0.6.0=pypi_0
defusedxml=0.7.1=pypi_0
diffusers=0.31.0=pypi_0
docker-pycreds=0.4.0=pypi_0
easydict=1.13=pypi_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pyhd8ed1ab_0
fastapi=0.115.6=pypi_0
ffmpy=0.4.0=pypi_0
filelock=3.16.1=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.10.0=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.37.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
gradio=4.44.1=pypi_0
gradio-client=1.3.0=pypi_0
grpcio=1.68.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.11.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.0=pypi_0
huggingface-hub=0.26.2=pypi_0
idna=3.10=pypi_0
imageio=2.19.3=pypi_0
imageio-ffmpeg=0.4.7=pypi_0
importlib-metadata=8.5.0=pyha770c72_0
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.5=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab-widgets=3.0.13=pypi_0
keras=2.13.1=pypi_0
kiwisolver=1.4.7=pypi_0
kornia=0.7.3=pypi_0
kornia-rs=0.1.7=pypi_0
krb5=1.21.3=h143b758_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libclang=18.1.1=pypi_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libsodium=1.0.20=h4ab18f5_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
lightning=2.3.3=pypi_0
lightning-utilities=0.11.8=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
more-itertools=10.5.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.3=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.6.77=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.4.0=hb9d3cd8_0
opt-einsum=3.4.0=pypi_0
orjson=3.10.12=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.0.3=pypi_0
parso=0.8.4=pyhd8ed1ab_0
peft=0.13.2=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.2=py38h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.48=pyha770c72_0
prompt_toolkit=3.0.48=hd8ed1ab_1
propcache=0.2.0=pypi_0
protobuf=4.25.5=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.4=pypi_0
python=3.8.20=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.19=pypi_0
python_abi=3.8=2_cp38
pytorch-lightning=2.4.0=pypi_0
pytz=2024.2=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
rasterio=1.3.11=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rich=13.9.4=pypi_0
rsa=4.9=pypi_0
ruff=0.8.1=pypi_0
safetensors=0.4.5=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.9.1=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
sentry-sdk=2.18.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.1.0=py38h06a4308_0
shapely=2.0.6=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
snuggs=1.4.7=pypi_0
spaces=0.30.4=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.41.3=pypi_0
sympy=1.13.3=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.13.1=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.6.7=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.3=pypi_0
tomlkit=0.12.0=pypi_0
torch=2.2.1=pypi_0
torchaudio=2.2.1=pypi_0
torchmetrics=1.5.2=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.17.1=pypi_0
tornado=6.4.1=py38hfb59056_0
tqdm=4.67.0=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.46.2=pypi_0
triton=2.2.0=pypi_0
typer=0.15.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
uvicorn=0.32.1=pypi_0
wandb=0.18.7=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.6=pypi_0
wheel=0.44.0=py38h06a4308_0
widgetsnbextension=4.0.13=pypi_0
wrapt=1.17.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.15.2=pypi_0
zeromq=4.3.5=h3b0a872_7
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
